Add gauges and location to place

+++ src/model/checkin/common.rs
@@ -10,7 +10,7 @@ pub struct Checkin {
     pub user_id: Uuid,
     pub start_timestamp: DateTime<Utc>,
     pub end_timestamp: DateTime<Utc>,
-    pub duration: i64,
+    pub duration: i32,
     pub potential_infection: bool,
     pub confirmed: bool,
     pub updated_at: DateTime<Utc>,
@@ -25,6 +25,6 @@ pub struct CheckinInsert {
     pub user_id: Uuid,
     pub start_timestamp: DateTime<Utc>,
     pub end_timestamp: DateTime<Utc>,
-    pub duration: i64,
+    pub duration: i32,
     pub confirmed: bool,
 }
