+++ src/serve/types/checkin.rs
@@ -13,7 +13,7 @@ pub struct CheckinForm {
     #[validate(email)]
     pub email: String,
     #[validate(range(min = 1, max = 480))]
-    pub duration: i64,
+    pub duration: i32,
 }
 
 #[derive(Serialize)]
@@ -22,7 +22,7 @@ pub struct Checkin {
     pub id: Uuid,
     pub start_timestamp: DateTime<Utc>,
     pub end_timestamp: DateTime<Utc>,
-    pub duration: i64,
+    pub duration: i32,
     pub potential_infection: bool,
     pub place: Place,
 }
